requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867338309/work
aiohttp @ file:///C:/ci/aiohttp_1614361133907/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831764494/work
chardet @ file:///C:/ci/chardet_1605303259695/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
coverage @ file:///C:/ci/coverage_1614614910274/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
cycler==0.10.0
Cython @ file:///C:/ci/cython_1626257059178/work
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1614884419385/work
h5py @ file:///C:/ci/h5py_1622088627516/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877486026/work
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keras2onnx @ git+https://github.com/onnx/keras-onnx@1d66c99b3ccc3e109bcf28c6475b4d3a5ca0015a
kiwisolver==1.3.1
Markdown @ file:///C:/ci/markdown_1614364005059/work
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1607367865052/work
numexpr @ file:///C:/ci/numexpr_1618856761305/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271861146/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
onnxconverter-common @ git+https://github.com/microsoft/onnxconverter-common@46a72b90d25bb92231c82c48a4c7828bbca167fe
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pandas @ file:///C:/ci/pandas_1626784212000/work
Pillow @ file:///C:/ci/pillow_1625663286921/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT @ file:///C:/ci/pyjwt_1619651810943/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///C:/ci/scipy_1618856134946/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorboard_1622025863084/work/tensorboard-2.5.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
tensorflow==2.5.0
tensorflow-estimator @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorflow-estimator_1622026529081/work/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
termcolor==1.1.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
